requirements.txt
fastapi[standard]
python-multipart 
langchain_community
langchain_text_splitters
langchain-google-genai
faiss-cpu 
pypdf 
unstructured 
python-docx
python-dotenv
pymongo[srv]
passlib[bcrypt] 
python-jose
pytest 
pytest-cov
coverage
